requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch==2.2.2
torchvision==0.17.2
transformers==4.40.0
huggingface_hub==0.22.2
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.0
pandas==2.2.2
